requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aiohttp==1.0.5            # via discord.py
async-timeout==2.0.1      # via aiohttp
beautifulsoup4==4.6.0     # via bs4
bs4==0.0.1
certifi==2018.4.16        # via requests
chardet==3.0.4            # via aiohttp, requests
configparser==3.5.0
datetime==4.2
discord.py==0.16.12
idna==2.6                 # via requests
imgurpython==1.1.7
inflect==0.2.5            # via jaraco.itertools
jaraco.itertools==2.2     # via wolframalpha
more-itertools==4.1.0     # via jaraco.itertools
multidict==4.2.0          # via aiohttp
praw==5.4.0
prawcore==0.14.0          # via praw
pytz==2018.4              # via datetime
requests==2.18.4          # via imgurpython, prawcore, update-checker
selenium==3.11.0
six==1.11.0               # via jaraco.itertools, more-itertools, wolframalpha
update-checker==0.16      # via praw
urllib3==1.22             # via requests
websockets==3.4           # via discord.py
wolframalpha==3.0.1
xmltodict==0.11.0         # via wolframalpha
zope.interface==4.5.0     # via datetime
